requirements.txt
Django==6.0.4
Pillow==12.0.0
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
python-dotenv==1.1.1
psycopg[binary]==3.2.3
